Conflicts:
	AUTHORS
	CHANGELOG
	INSTALL
	Makefile
	SConstruct
	agg/SConscript
	agg/build.py
	bindings/python/build.py
	bindings/python/mapnik/__init__.py
	bindings/python/mapnik_feature.cpp
	bindings/python/mapnik_glyph_symbolizer.cpp
	bindings/python/mapnik_image.cpp
	bindings/python/mapnik_image_view.cpp
	bindings/python/mapnik_palette.cpp
	bindings/python/mapnik_polygon_pattern_symbolizer.cpp
	bindings/python/mapnik_python.cpp
	bindings/python/mapnik_rule.cpp
	bindings/python/mapnik_shield_symbolizer.cpp
	bindings/python/mapnik_text_symbolizer.cpp
	bindings/python/mapnik_value_converter.hpp
	demo/c++/build.py
	deps/agg/build.py
	include/mapnik/agg_renderer.hpp
	include/mapnik/attribute_collector.hpp
	include/mapnik/box2d.hpp
	include/mapnik/cairo_renderer.hpp
	include/mapnik/color.hpp
	include/mapnik/expression_evaluator.hpp
	include/mapnik/feature_layer_desc.hpp
	include/mapnik/feature_style_processor.hpp
	include/mapnik/font_engine_freetype.hpp
	include/mapnik/glyph_symbolizer.hpp
	include/mapnik/grid/grid.hpp
	include/mapnik/grid/grid_renderer.hpp
	include/mapnik/hextree.hpp
	include/mapnik/image_util.hpp
	include/mapnik/label_placement.hpp
	include/mapnik/load_map.hpp
	include/mapnik/palette.hpp
	include/mapnik/path_expression_grammar.hpp
	include/mapnik/placement_finder.hpp
	include/mapnik/png_io.hpp
	include/mapnik/rule.hpp
	include/mapnik/sql_utils.hpp
	include/mapnik/text_path.hpp
	include/mapnik/text_placements.hpp
	include/mapnik/text_symbolizer.hpp
	include/mapnik/timer.hpp
	include/mapnik/value.hpp
	osx/config.py
	osx/installer/scripts/setup.sh
	osx/scripts/fetch_py_releases.sh
	osx/scripts/package.py
	osx/scripts/setup.sh
	plugins/input/gdal/build.py
	plugins/input/geos/build.py
	plugins/input/kismet/build.py
	plugins/input/occi/build.py
	plugins/input/ogr/build.py
	plugins/input/osm/build.py
	plugins/input/postgis/build.py
	plugins/input/postgis/postgis_datasource.cpp
	plugins/input/postgis/postgis_datasource.hpp
	plugins/input/postgis/postgis_featureset.cpp
	plugins/input/postgis/postgis_featureset.hpp
	plugins/input/raster/build.py
	plugins/input/rasterlite/build.py
	plugins/input/shape/build.py
	plugins/input/shape/shape_datasource.cpp
	plugins/input/shape/shape_featureset.cpp
	plugins/input/shape/shape_featureset.hpp
	plugins/input/shape/shape_index_featureset.cpp
	plugins/input/shape/shape_index_featureset.hpp
	plugins/input/shape/shape_io.cpp
	plugins/input/shape/shape_io.hpp
	plugins/input/sqlite/build.py
	plugins/input/sqlite/sqlite_datasource.cpp
	plugins/input/sqlite/sqlite_datasource.hpp
	plugins/input/sqlite/sqlite_featureset.cpp
	plugins/input/sqlite/sqlite_featureset.hpp
	plugins/input/sqlite/sqlite_resultset.hpp
	plugins/input/templates/helloworld/build.py
	scons/scons-LICENSE
	scons/scons-README
	scons/scons-time.py
	scons/scons.py
	scons/sconsign.py
	src/agg/process_building_symbolizer.cpp
	src/agg/process_glyph_symbolizer.cpp
	src/agg/process_point_symbolizer.cpp
	src/agg/process_polygon_pattern_symbolizer.cpp
	src/agg/process_raster_symbolizer.cpp
	src/agg/process_shield_symbolizer.cpp
	src/agg/process_text_symbolizer.cpp
	src/build.py
	src/cairo_renderer.cpp
	src/datasource_cache.cpp
	src/feature_style_processor.cpp
	src/font_engine_freetype.cpp
	src/glyph_symbolizer.cpp
	src/grid/process_building_symbolizer.cpp
	src/grid/process_point_symbolizer.cpp
	src/grid/process_text_symbolizer.cpp
	src/image_util.cpp
	src/layer.cpp
	src/load_map.cpp
	src/metawriter.cpp
	src/palette.cpp
	src/placement_finder.cpp
	src/save_map.cpp
	src/svg/process_glyph_symbolizer.cpp
	src/text_placements.cpp
	src/text_symbolizer.cpp
	tests/cpp_tests/build.py
	tests/cpp_tests/svg_renderer_tests/build.py
	tests/data/good_maps/also_and_else_filter.xml
	tests/data/good_maps/glyph_symbolizer.xml
	tests/data/good_maps/sqlite_attachdb.xml
	tests/data/sqlite/world.sqlite
	tests/python_tests/datasource_test.py
	tests/python_tests/feature_id_test.py
	tests/python_tests/feature_test.py
	tests/python_tests/filter_test.py
	tests/python_tests/font_test.py
	tests/python_tests/glyph_symbolizer_test.py
	tests/python_tests/introspection_test.py
	tests/python_tests/layer_modification_test.py
	tests/python_tests/object_test.py
	tests/python_tests/raster_colorizer_test.py
	tests/python_tests/render_grid_test.py
	tests/python_tests/render_test.py
	tests/python_tests/sqlite_test.py
	tests/run_tests.py
	utils/mapnik-config/build.py
	utils/mapnik-config/mapnik-config.template.sh
	utils/ogrindex/build.py
	utils/performance/build.py
	utils/pgsql2sqlite/build.py
	utils/shapeindex/build.py
	utils/shapeindex/shapeindex.cpp
	utils/svg2png/build.py
	utils/upgrade_map_xml/upgrade_map_xml.py

+++ src/shield_symbolizer.cpp
@@ -2,8 +2,7 @@
  *
  * This file is part of Mapnik (c++ mapping toolkit)
  *
- * Copyright (C) 2006 Artem Pavlenko
- * Copyright (C) 2006 10East Corp.
+ * Copyright (C) 2011 Artem Pavlenko
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -47,7 +46,7 @@ shield_symbolizer::shield_symbolizer(text_placements_ptr placements)
 shield_symbolizer::shield_symbolizer(
     expression_ptr name,
     std::string const& face_name,
-    unsigned size,
+    float size,
     color const& fill,
     path_expression_ptr file)
     : text_symbolizer(name, face_name, size, fill),
@@ -60,7 +59,7 @@ shield_symbolizer::shield_symbolizer(
 
 shield_symbolizer::shield_symbolizer(
     expression_ptr name,
-    unsigned size,
+    float size,
     color const& fill,
     path_expression_ptr file)
     : text_symbolizer(name, size, fill),
